esercizio 33

+++ src/Esercizio33/Forma.java
@@ -1,7 +1,7 @@
 package Esercizio33;
 
 public class Forma {
-    public int calcolaArea() {
+    public double calcolaArea() {
 
         System.out.println("La forma è: "+"Rettangolo");
         return 0;
